# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Aikya Sentinel – Unified AI Financial Crime Intelligence & Insider Threat Early Warning System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /accounts:
    get:
      operationId: listAccounts
      tags: [accounts]
      summary: List accounts
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Account"
    post:
      operationId: createAccount
      tags: [accounts]
      summary: Create account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
  /accounts/{id}:
    get:
      operationId: getAccount
      tags: [accounts]
      summary: Get account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "404":
          description: Not found
    patch:
      operationId: updateAccount
      tags: [accounts]
      summary: Update account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
    delete:
      operationId: deleteAccount
      tags: [accounts]
      summary: Delete account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: No content

  # Transactions
...
  /behaviour/analyse:
    post:
      operationId: analyseBehaviour
      tags: [ai]
      summary: M1 - Behaviour DNA Engine
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EntityAnalysisInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BehaviourAnalysisResult"

  # M2 - Insider Threat Engine
  /insider/analyse:
    post:
      operationId: analyseInsiderThreat
      tags: [ai]
      summary: M2 - Insider Threat Engine
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EntityAnalysisInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InsiderThreatResult"

  # M3 - Fund Flow Intelligence
  /fundflow/analyse:
    post:
      operationId: analyseFundFlow
      tags: [ai]
      summary: M3 - Fund Flow Intelligence
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FundFlowInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FundFlowResult"

  # M4 - Lifestyle vs Income Risk
  /taxrisk/analyse:
    post:
      operationId: analyseTaxRisk
      tags: [ai]
      summary: M4 - Lifestyle vs Income Risk
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EntityAnalysisInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaxRiskResult"

  # M5 - Shell/Mule Detection
  /shell/analyse:
    post:
      operationId: analyseShellMule
      tags: [ai]
      summary: M5 - Shell/Mule Detection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EntityAnalysisInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShellMuleResult"

  # M6 - Collusion Detection
  /collusion/analyse:
    post:
      operationId: analyseCollusion
      tags: [ai]
      summary: M6 - Collusion Detection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CollusionInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollusionResult"

  # M7 - Fraud Intent Prediction
  /intent/analyse:
    post:
      operationId: analyseFraudIntent
      tags: [ai]
      summary: M7 - Fraud Intent Prediction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EntityAnalysisInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FraudIntentResult"

  # M8 - Digital Behaviour Twin
  /twin/analyse:
    post:
      operationId: analyseDigitalTwin
      tags: [ai]
      summary: M8 - Digital Behaviour Twin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EntityAnalysisInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DigitalTwinResult"

  # M9 - Pattern Genome
  /patterns/match:
    post:
      operationId: matchPatterns
      tags: [ai]
      summary: M9 - Pattern Genome
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatternMatchInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatternMatchResult"

  # M10 - Risk Radar
  /riskradar/analyse:
    post:
      operationId: analyseRiskRadar
      tags: [ai]
      summary: M10 - Risk Radar (aggregate all module scores)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EntityAnalysisInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RiskRadarResult"

  # M11 - Explainable AI
  /xai/explain:
    post:
      operationId: explainDecision
      tags: [ai]
      summary: M11 - Explainable AI (SHAP explanations)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/XAIInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/XAIResult"

  # M12 - Timeline Tracker
  /timeline/entity/{entityId}:
    get:
      operationId: fetchEntityTimeline
      tags: [ai]
      summary: M12 - Timeline Tracker
      parameters:
        - name: entityId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimelineResult"

  # M13 - Trust Score Engine
  /trust/calculate:
    post:
      operationId: calculateTrustScore
      tags: [ai]
      summary: M13 - Trust Score Engine
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EntityAnalysisInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrustScoreResult"

  # M15 - Investigation Narrative Generator
  /reports/generate:
    post:
      operationId: generateReport
      tags: [ai]
      summary: M15 - Investigation Narrative Generator
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportResult"
...
  /copilot/chat:
    post:
      operationId: copilotChat
      tags: [ai]
      summary: M16 - Investigator Copilot chat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CopilotChatInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CopilotChatResult"

  # M17 - Fraud Simulation Mode
  /simulation/run:
    post:
      operationId: runSimulation
      tags: [ai]
      summary: M17 - Fraud Simulation Mode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SimulationInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimulationResult"

  # M18 - Risk Personality Profiling
  /personality/analyse:
    post:
      operationId: analysePersonality
      tags: [ai]
      summary: M18 - Risk Personality Profiling
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EntityAnalysisInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PersonalityResult"

  # M19 - Adaptive Learning
  /learning/feedback:
    post:
      operationId: submitLearningFeedback
      tags: [ai]
      summary: M19 - Adaptive Learning feedback
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LearningFeedbackInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # M20 - Risk Heatmap Dashboard API
...
  /risk-scores/{entityId}:
    get:
      operationId: getRiskScores
      tags: [ai]
      summary: Get composite risk scores for an entity
      parameters:
        - name: entityId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RiskScore"
components:
  schemas:
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
...
    Account:
      type: object
      required: [id, accountId, customerName, kycStatus, riskScore, createdAt]
      properties:
        id:
          type: integer
        accountId:
          type: string
        customerName:
          type: string
        kycStatus:
          type: string
        riskScore:
          type: number
        accountType:
          type: ["string", "null"]
        balance:
          type: ["number", "null"]
        createdAt:
          type: string
    AccountInput:
      type: object
      required: [accountId, customerName, kycStatus]
      properties:
        accountId:
          type: string
        customerName:
          type: string
        kycStatus:
          type: string
        riskScore:
          type: number
        accountType:
          type: string
        balance:
          type: number
    AccountUpdate:
      type: object
      properties:
        customerName:
          type: string
        kycStatus:
          type: string
        riskScore:
          type: number
        accountType:
          type: string
...
    EntityAnalysisInput:
      type: object
      required: [entityId, entityType]
      properties:
        entityId:
          type: string
        entityType:
          type: string
          enum: [employee, account, transaction]
    BehaviourAnalysisResult:
      type: object
      required: [entityId, behaviourScore, riskLevel, signals, analysedAt]
      properties:
        entityId:
          type: string
        behaviourScore:
          type: number
        riskLevel:
          type: string
          enum: [LOW, MEDIUM, HIGH, CRITICAL]
        signals:
          type: array
          items:
            type: string
        anomalies:
          type: array
          items:
            type: string
        analysedAt:
          type: string
    InsiderThreatResult:
      type: object
      required: [entityId, insiderRisk, riskLevel, indicators, analysedAt]
      properties:
        entityId:
          type: string
        insiderRisk:
          type: number
        riskLevel:
          type: string
        indicators:
          type: array
          items:
            type: string
        analysedAt:
          type: string
    FundFlowInput:
      type: object
      required: [accountId]
      properties:
        accountId:
          type: string
        depth:
          type: integer
    FundFlowResult:
      type: object
      required: [accountId, patterns, nodes, edges, analysedAt]
      properties:
        accountId:
          type: string
        patterns:
          type: array
          items:
            $ref: "#/components/schemas/FundFlowPattern"
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/GraphNode"
        edges:
          type: array
          items:
            $ref: "#/components/schemas/GraphEdge"
        analysedAt:
          type: string
    FundFlowPattern:
      type: object
      required: [type, detected, confidence]
      properties:
        type:
          type: string
          enum: [circular_routing, layering, structuring, dormant_activation]
        detected:
          type: boolean
        confidence:
          type: number
        description:
          type: string
    GraphNode:
      type: object
      required: [id, label, type]
      properties:
        id:
          type: string
        label:
          type: string
        type:
          type: string
        riskScore:
          type: ["number", "null"]
    GraphEdge:
      type: object
      required: [source, target, amount]
      properties:
        source:
          type: string
        target:
          type: string
        amount:
          type: number
        timestamp:
          type: ["string", "null"]
    TaxRiskResult:
      type: object
      required: [entityId, taxRiskScore, riskLevel, lifestyle, reportedIncome, gap, analysedAt]
      properties:
        entityId:
          type: string
        taxRiskScore:
          type: number
        riskLevel:
          type: string
        lifestyle:
          type: number
        reportedIncome:
          type: number
        gap:
          type: number
        flags:
          type: array
          items:
            type: string
        analysedAt:
          type: string
    ShellMuleResult:
      type: object
      required: [entityId, isShell, isMule, confidence, indicators, analysedAt]
      properties:
        entityId:
          type: string
        isShell:
          type: boolean
        isMule:
          type: boolean
        confidence:
          type: number
        indicators:
          type: array
          items:
            type: string
        analysedAt:
          type: string
    CollusionInput:
      type: object
      required: [entityIds]
      properties:
        entityIds:
          type: array
          items:
            type: string
    CollusionResult:
      type: object
      required: [collusionDetected, confidence, clusters, analysedAt]
      properties:
        collusionDetected:
          type: boolean
        confidence:
          type: number
        clusters:
          type: array
          items:
            $ref: "#/components/schemas/CollusionCluster"
        analysedAt:
          type: string
    CollusionCluster:
      type: object
      required: [members, riskScore, evidence]
      properties:
        members:
          type: array
          items:
            type: string
        riskScore:
          type: number
        evidence:
          type: array
          items:
            type: string
    FraudIntentResult:
      type: object
      required: [entityId, fraudIntentScore, riskLevel, prediction, confidence, analysedAt]
      properties:
        entityId:
          type: string
        fraudIntentScore:
          type: number
        riskLevel:
          type: string
        prediction:
          type: string
        confidence:
          type: number
        factors:
          type: array
          items:
            type: string
        analysedAt:
          type: string
    DigitalTwinResult:
      type: object
      required: [entityId, twinProfile, deviations, anomalyScore, analysedAt]
      properties:
        entityId:
          type: string
        twinProfile:
          type: object
          additionalProperties: {}
        deviations:
          type: array
          items:
            type: string
        anomalyScore:
          type: number
        analysedAt:
          type: string
    PatternMatchInput:
      type: object
      required: [entityId, patternTypes]
      properties:
        entityId:
          type: string
        patternTypes:
          type: array
          items:
            type: string
    PatternMatchResult:
      type: object
      required: [entityId, matchedPatterns, genomeSignature, analysedAt]
      properties:
        entityId:
          type: string
        matchedPatterns:
          type: array
          items:
            $ref: "#/components/schemas/PatternMatch"
        genomeSignature:
          type: string
        analysedAt:
          type: string
    PatternMatch:
      type: object
      required: [patternId, name, confidence, severity]
      properties:
        patternId:
          type: string
        name:
          type: string
        confidence:
          type: number
        severity:
          type: string
    RiskRadarResult:
      type: object
      required: [entityId, fraudScore, amlScore, behaviourScore, taxScore, networkScore, overallScore, tier, analysedAt]
      properties:
        entityId:
          type: string
        fraudScore:
          type: number
        amlScore:
          type: number
        behaviourScore:
          type: number
        taxScore:
          type: number
        networkScore:
          type: number
        overallScore:
          type: number
        tier:
          type: string
          enum: [GREEN, AMBER, RED, BLACK]
        analysedAt:
          type: string
    XAIInput:
      type: object
      required: [entityId, modelType]
      properties:
        entityId:
          type: string
        modelType:
          type: string
    XAIResult:
      type: object
      required: [entityId, modelType, explanation, shapValues, topFeatures, analysedAt]
      properties:
        entityId:
          type: string
        modelType:
          type: string
        explanation:
          type: string
        shapValues:
          type: object
          additionalProperties:
            type: number
        topFeatures:
          type: array
          items:
            $ref: "#/components/schemas/FeatureImportance"
        analysedAt:
          type: string
    FeatureImportance:
      type: object
      required: [feature, importance, direction]
      properties:
        feature:
          type: string
        importance:
          type: number
        direction:
          type: string
          enum: [positive, negative]
    TimelineResult:
      type: object
      required: [entityId, entityType, events, analysedAt]
      properties:
        entityId:
          type: string
        entityType:
          type: string
        events:
          type: array
          items:
            $ref: "#/components/schemas/TimelineEvent"
        analysedAt:
          type: string
    TimelineEvent:
      type: object
      required: [id, eventType, description, timestamp, riskImpact]
      properties:
        id:
          type: string
        eventType:
          type: string
        description:
          type: string
        timestamp:
          type: string
        riskImpact:
          type: string
          enum: [LOW, MEDIUM, HIGH, CRITICAL]
        metadata:
          type: object
          additionalProperties: {}
    TrustScoreResult:
      type: object
      required: [entityId, trustScore, trustLevel, components, analysedAt]
      properties:
        entityId:
          type: string
        trustScore:
          type: number
        trustLevel:
          type: string
          enum: [TRUSTED, MODERATE, SUSPICIOUS, UNTRUSTED]
        components:
          type: object
          additionalProperties:
            type: number
        analysedAt:
          type: string
    ReportInput:
      type: object
      required: [entityId, entityType, reportType]
      properties:
        entityId:
          type: string
        entityType:
          type: string
        reportType:
          type: string
          enum: [investigation, sar, aml_summary, risk_profile]
    ReportResult:
      type: object
      required: [reportId, entityId, reportType, narrative, findings, recommendations, generatedAt]
      properties:
        reportId:
          type: string
        entityId:
          type: string
        reportType:
          type: string
        narrative:
          type: string
        findings:
          type: array
          items:
            type: string
        recommendations:
          type: array
          items:
            type: string
        riskLevel:
          type: ["string", "null"]
        generatedAt:
          type: string
    CopilotChatInput:
      type: object
      required: [message]
      properties:
        message:
          type: string
        sessionId:
          type: ["string", "null"]
        context:
          type: ["string", "null"]
    CopilotChatResult:
      type: object
      required: [response, sessionId, sources]
      properties:
        response:
          type: string
        sessionId:
          type: string
        sources:
          type: array
          items:
            type: string
        suggestedActions:
          type: array
          items:
            type: string
    SimulationInput:
      type: object
      required: [scenario]
      properties:
        scenario:
          type: string
          enum: [aml, insider_threat, collusion]
        parameters:
          type: object
          additionalProperties: {}
    SimulationResult:
      type: object
      required: [scenario, simulationId, events, alertsTriggered, riskScore, summary, simulatedAt]
      properties:
        scenario:
          type: string
        simulationId:
          type: string
        events:
          type: array
          items:
            $ref: "#/components/schemas/SimulationEvent"
        alertsTriggered:
          type: integer
        riskScore:
          type: number
        summary:
          type: string
        simulatedAt:
          type: string
...
    PersonalityResult:
      type: object
      required: [entityId, personalityType, traits, riskProfile, analysedAt]
      properties:
        entityId:
          type: string
        personalityType:
          type: string
        traits:
          type: array
          items:
            $ref: "#/components/schemas/PersonalityTrait"
        riskProfile:
          type: string
        analysedAt:
          type: string
...
    LearningFeedbackInput:
      type: object
      required: [entityId, modelType, prediction, actualOutcome]
      properties:
        entityId:
          type: string
        modelType:
          type: string
        prediction:
          type: string
        actualOutcome:
          type: string
        feedback:
          type: ["string", "null"]
    RiskScore:
      type: object
      required: [entityId, fraudScore, amlScore, behaviourScore, taxScore, networkScore, overallScore, updatedAt]
      properties:
        entityId:
          type: string
        fraudScore:
          type: number
        amlScore:
          type: number
        behaviourScore:
          type: number
        taxScore:
          type: number
        networkScore:
          type: number
        overallScore:
          type: number
        updatedAt:
          type: string
...